Export GO111MODULE environment variable to turn on module mode
Module mode was introduced in go 1.11 and will become the default in go 1.13

Add '-mod=vendor' to 'go build' and 'go install' invocations via GOFLAGS
variable.

In go1.11 and higher, when building in module mode,
this instructs the toolchain to look in the top-level
vendor directory for modules instead of downloading them.

Signed-off-by: Craig Rodrigues <craig@portworx.com>